requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
edge-tts==6.1.9
google-generativeai==0.5.4
faster-whisper==1.0.3